requirements.txt
telethon
python-telegram-bot
requests